requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-filter>=23.5
psycopg2-binary>=2.9.9
Pillow>=10.1.0
python-dotenv>=1.0.0
celery>=5.3.6
redis>=5.0.1
djangorestframework-simplejwt>=5.3.1
django-celery-beat>=2.5.0
django-celery-results>=2.5.1
gunicorn>=21.2.0
whitenoise>=6.6.0
django-storages>=1.14.2
boto3>=1.34.34
requests>=2.31.0
python-dateutil>=2.8.2
pytz>=2024.1 